backend/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==3.1.0
psycopg2-binary==2.9.3
redis==4.3.4
python-dotenv==0.20.0
Flask-Cors==3.0.10
PyJWT==2.4.0
werkzeug==2.2.2
SQLAlchemy==1.4.39

